Fix self-signed certificate error from Google
There was code to create our own PEM for autentication. This failed at Google's IdP because the certificate is self-signed.
