+++ .vscode/launch.json
@@ -5,6 +5,15 @@
     "version": "0.2.0",
     "configurations": [
         {
+            "type": "node",
+            "request": "attach",
+            "protocol": "inspector",
+            "name": "Launch API",
+            "port": 9229,
+            "sourceMaps": true,
+            "trace": true
+         },
+        {
             "name": "PowerShell Interactive Session",
             "type": "PowerShell",
             "request": "launch",
