+++ EdFi.Buzz.Api/package.json
@@ -39,6 +39,7 @@
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-react": "^7.20.5",
     "eslint-plugin-react-hooks": "^4.0.8",
+    "google-auth-library": "^6.1.3",
     "graphile-worker": "^0.6.0",
     "graphql": "14.5.0",
     "graphql-tools": "6.0.9",
